fixing ci test errors
